clickable label to the right place
